// src/app/ledger-previous-trans-core/models/enums.ts
export enum AllGridColumnWidth {
    Date = '80px',
    Type = '85px',
    URN = '85px',
    Ref = '90px',
    Details = '',
    // VAT = '85px',
    // Period = '80px',
    OfficeDr = '95px',
    OfficeCr = '95px',
    OfficeBal = '98px',
    ClientDr = '115px',
    ClientCr = '115px',
    ClientBal = '115px',
    DDADr = '95px',
    DDDCr = '95px',
    DDDbal = '95px'
}
// account: string;
// clientBal: number;
// clientCr ?: number;
// clientDr ?: number;
// date: string;
// ddDbal: number;
// ddaDr ?: number;
// ddaCr ?: number;
// details: string;
// disputed ?: number;
// feeEarner ?: string;
// fullDetails ?: string;
// gross ?: number;
// id ?: string;
// net ?: number;
// officeBal: number;
// officeCr ?: number;
// officeDr: number;
// period ?: number;
// ptId: number;
// ref: string;
// reverse ?: number;
// type: string;
// uiType: number;
// urn ?: number;
// vat ?: number;
export enum GridFilterType {
    showAll = 'showAll',
    showOffice = 'showOffice',
    showClient = 'showClient'
}